requirements.txt
Flask
flask_sqlalchemy 
flask_jwt_extended
flask_migrate
flask_cors
flask_restful
sqlalchemy_serializer 
flask_bcrypt
python-dotenv